Rename coordinality with cardinality.
