+++ Modules/Migration/ViewModels/IFixCardinalityErrorsDialogViewModel.cs
@@ -1,0 +1,38 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="IFixCardinalityErrorsDialogViewModel.cs" company="RHEA System S.A.">
+//    Copyright (c) 2015-2020 RHEA System S.A.
+//
+//    Author: Adrian Chivu, Cozmin Velciu, Alex Vorobiev
+//
+//    This file is part of CDP4-Server-Administration-Tool.
+//    The CDP4-Server-Administration-Tool is an ECSS-E-TM-10-25 Compliant tool
+//    for advanced server administration.
+//
+//    The CDP4-Server-Administration-Tool is free software; you can redistribute it and/or modify
+//    it under the terms of the GNU Affero General Public License as
+//    published by the Free Software Foundation; either version 3 of the
+//    License, or (at your option) any later version.
+//
+//    The CDP4-Server-Administration-Tool is distributed in the hope that it will be useful,
+//    but WITHOUT ANY WARRANTY; without even the implied warranty of
+//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+//    Affero General Public License for more details.
+//
+//    You should have received a copy of the GNU Affero General Public License
+//    along with this program. If not, see <http://www.gnu.org/licenses/>.
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace Migration.ViewModels
+{
+    /// <summary>
+    /// The interface for the view model <see cref="FixCardinalityErrorsDialogViewModel"/>
+    /// </summary>
+    public interface IFixCardinalityErrorsDialogViewModel
+    {
+        /// <summary>
+        /// Apply PocoCardinality & PocoProperties to the E10-25 data set and bind errors to the reactive list
+        /// </summary>
+        void BindPocoErrors();
+    }
+}
