+++ Modules/Migration/Views/FixCardinalityErrorsDialog.xaml
@@ -1,0 +1,68 @@
+﻿<dx:ThemedWindow
+    x:Class="Migration.Views.FixCardinalityErrorsDialog"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:dx="http://schemas.devexpress.com/winfx/2008/xaml/core"
+    xmlns:dxg="http://schemas.devexpress.com/winfx/2008/xaml/grid"
+    xmlns:dxe="http://schemas.devexpress.com/winfx/2008/xaml/editors"
+    WindowStartupLocation="CenterScreen"
+    Title="Cardinality Errors" Height="600" Width="1000" Loaded="FixCardinalityErrorsDialog_OnLoaded"
+    Icon="pack://application:,,,/SAT;component/Resources/cdplogo3d_48x48.png">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="65"/>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="130"></RowDefinition>
+            <RowDefinition Height="65"/>
+        </Grid.RowDefinitions>
+        <TextBlock TextWrapping="WrapWithOverflow" Padding="20" Grid.Row="0">
+            The following table summarizes the errors found in the data selected for migration. You can attempt to fix it locally before sending it off to the target server. Failure to do so might result in the target server rejecting the migration package.
+        </TextBlock>
+        <dxg:GridControl Grid.Row="1" Margin="20"
+            AllowLiveDataShaping="False"
+            ShowLoadingPanel="{Binding IsBusy, UpdateSourceTrigger=PropertyChanged}"
+            SelectedItem="{Binding SelectedError, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
+            SelectionMode="Row"
+            ItemsSource="{Binding Errors, UpdateSourceTrigger=PropertyChanged}"
+            AutoGenerateColumns="None">
+            <dxg:GridControl.View>
+                <dxg:TableView
+                    ShowIndicator="False"
+                    HorizontalAlignment="Stretch"
+                    VerticalAlignment="Stretch"
+                    AllowColumnMoving="False"
+                    AllowEditing="False"
+                    AllowGrouping="False"
+                    AutoWidth="True"
+                    ShowGroupPanel="False"
+                    IsDetailButtonVisibleBinding="{x:Null}"
+                    VerticalScrollbarVisibility="Auto">
+                </dxg:TableView>
+            </dxg:GridControl.View>
+            <dxg:GridControl.Columns>
+                <dxg:GridColumn AllowColumnFiltering="False" FieldName="TopContainerName" Header="Top Container" />
+                <dxg:GridColumn AllowGrouping="True" FieldName="ContainerThingClassKind" Header="Class Kind" />
+                <dxg:GridColumn AllowColumnFiltering="False" FieldName="Error" Header="Content" />
+                <dxg:GridColumn AllowColumnFiltering="False" FieldName="Path" Header="Path" />
+            </dxg:GridControl.Columns>
+        </dxg:GridControl>
+        <GroupBox Margin="20 0 20 0" Grid.Row="2" Header="Error details">
+            <dxe:TextEdit
+                Margin="0 0 0 0"
+                ShowBorder="False"
+                Padding="0"
+                FontSize="12"
+                TextWrapping="Wrap"
+                AcceptsReturn="True"
+                VerticalContentAlignment="Top"
+                VerticalScrollBarVisibility="Visible"
+                IsReadOnly="True"
+                Text="{Binding Path=ErrorDetails, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" />
+        </GroupBox>
+    </Grid>
+    <dx:ThemedWindow.DialogButtons>
+        <dx:ThemedWindowDialogButton Margin="0 0 20 0"  Content="Migrate" IsCancel="False" IsDefault="True" DialogResult="OK"></dx:ThemedWindowDialogButton>
+        <dx:ThemedWindowDialogButton Content="Cancel" IsCancel="True" DialogResult="Cancel"></dx:ThemedWindowDialogButton>
+        <dx:ThemedWindowDialogButton Content="Attempt Fix" IsCancel="False" Alignment="Left" Command="{Binding FixCommand}" AllowCloseWindow="False" DialogResult="None"></dx:ThemedWindowDialogButton>
+    </dx:ThemedWindow.DialogButtons>
+</dx:ThemedWindow>
